# API/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Aim/{aimId}/sources/{sourceId}": {
      "post": {
        "tags": [
          "Aim"
        ],
        "parameters": [
          {
            "name": "aimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Aim"
        ],
        "parameters": [
          {
            "name": "aimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/PlannedTransaction": {
      "get": {
        "tags": [
          "PlannedTransaction"
        ],
        "parameters": [
          {
            "name": "Offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "MinAmount",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "MaxAmount",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "SortDescending",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlannedTransactionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlannedTransactionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlannedTransactionDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PlannedTransaction"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlannedTransactionInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlannedTransactionInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlannedTransactionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedTransactionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedTransactionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedTransactionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/PlannedTransaction/{id}": {
      "get": {
        "tags": [
          "PlannedTransaction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedTransactionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedTransactionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedTransactionDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PlannedTransaction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "PlannedTransaction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlannedTransactionInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlannedTransactionInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlannedTransactionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedTransactionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedTransactionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedTransactionDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Source": {
      "get": {
        "tags": [
          "Source"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceDtoLookup"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceDtoLookup"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceDtoLookup"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Source"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSourceInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSourceInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSourceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              }
            }
          }
        }
      }
    },
    "/api/Source/{id}": {
      "get": {
        "tags": [
          "Source"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoDetailed"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoDetailed"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoDetailed"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Source"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSourceInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSourceInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSourceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              }
            }
          }
        }
      }
    },
    "/api/Source/archive/{id}": {
      "patch": {
        "tags": [
          "Source"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              }
            }
          }
        }
      }
    },
    "/api/Source/unarchive/{id}": {
      "patch": {
        "tags": [
          "Source"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Transaction/{id}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AimDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "isClosed": {
            "type": "boolean"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceDtoLookup"
            },
            "nullable": true
          },
          "progress": {
            "$ref": "#/components/schemas/AimProgressDto"
          }
        },
        "additionalProperties": false
      },
      "AimProgressDto": {
        "type": "object",
        "properties": {
          "collectedAmount": {
            "type": "number",
            "format": "double"
          },
          "targetAmount": {
            "type": "number",
            "format": "double"
          },
          "completionPercentage": {
            "type": "number",
            "format": "double",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CreatePlannedTransactionInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "currencyId": {
            "type": "integer",
            "format": "int32"
          },
          "transactionTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sourceId": {
            "type": "integer",
            "format": "int32"
          },
          "frequencyId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreateSourceInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currencyId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CurrencyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "usdExchangeRate": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "FrequencyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "intervalValue": {
            "type": "integer",
            "format": "int32"
          },
          "intervalUnit": {
            "$ref": "#/components/schemas/IntervalUnitDto"
          }
        },
        "additionalProperties": false
      },
      "IntervalUnitDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PlannedTransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryDto"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "frequency": {
            "$ref": "#/components/schemas/FrequencyDto"
          },
          "source": {
            "$ref": "#/components/schemas/SourceDtoLookup"
          },
          "transactionType": {
            "$ref": "#/components/schemas/TransactionTypeDto"
          }
        },
        "additionalProperties": false
      },
      "SourceDtoDetailed": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "isArchived": {
            "type": "boolean"
          },
          "aims": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AimDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SourceDtoLookup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "isArchived": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "TransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryDto"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "destinationSource": {
            "$ref": "#/components/schemas/SourceDtoLookup"
          },
          "source": {
            "$ref": "#/components/schemas/SourceDtoLookup"
          },
          "transactionType": {
            "$ref": "#/components/schemas/TransactionTypeDto"
          }
        },
        "additionalProperties": false
      },
...
      "TransactionTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdatePlannedTransactionInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "currencyId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "transactionTypeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "categoryId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sourceId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "frequencyId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateSourceInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateTransactionInput": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "categoryId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sourceId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "destinationSourceId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "currencyId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "transactionTypeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}